sometimes react complains about invalid intervals. Fixed.
React sometimes logs:

"react.inc.js:19500 Warning: `value` prop on `input` should not be null. Consider using the empty string to clear the component or `undefined` for uncontrolled components."

This happens when the client / server clocks aren't perfectly synchronized, and a device's timestamps are in the future according to the client. I think this is a clj-time defect (an interval with start > end should be negative, not an error) but in any case, this commit works around the problem.